Break up opds for distributors as well
