+++ src/palace/manager/celery/tasks/playtime_entries.py
@@ -18,10 +18,10 @@ from sqlalchemy.sql.functions import coalesce, count, max as sql_max, sum
 
 from palace.manager.api.config import Configuration
 from palace.manager.celery.task import Task
-from palace.manager.integration.license.opds.odl.api import OPDS2WithODLApi
-from palace.manager.integration.license.opds.opds_for_distributors import (
+from palace.manager.integration.license.opds.for_distributors.api import (
     OPDSForDistributorsAPI,
 )
+from palace.manager.integration.license.opds.odl.api import OPDS2WithODLApi
 from palace.manager.service.celery.celery import QueueNames
 from palace.manager.sqlalchemy.model.collection import Collection
 from palace.manager.sqlalchemy.model.identifier import Identifier
